Rename interval utils and models
