+++ include/data_structures/shared_interval_model.h
