+++ src/mif/gavril.cpp
@@ -1,5 +1,5 @@
 
-#include "data_structures/distinct_interval_rep.h"
+#include "data_structures/distinct_interval_model.h"
 #include "data_structures/interval.h"
 
 #include "mif/gavril.h"
