+++ src/mis/distinct/implicit_output_sensitive.cpp
@@ -7,7 +7,7 @@
 
 #include <algorithm>
 
-#include "data_structures/distinct_interval_rep.h"
+#include "data_structures/distinct_interval_model.h"
 #include "data_structures/interval.h"
 #include "mis/implicit_independent_set.h"
 #include "mis/unit_monotone_seq.h"
