+++ src/mis/distinct/switching.cpp
@@ -1,7 +1,7 @@
-#include "data_structures/distinct_interval_rep.h"
+#include "data_structures/distinct_interval_model.h"
 #include "mis/distinct/valiente.h"
 #include "mis/distinct/pure_output_sensitive.h"
-#include "utils/interval_rep_utils.h"
+#include "utils/interval_model_utils.h"
 #include "utils/counters.h"
 #include "data_structures/interval.h"
 
@@ -12,7 +12,7 @@ namespace cg::mis::distinct
 {
     std::vector<cg::data_structures::Interval> Switching::computeMIS(const cg::data_structures::DistinctIntervalModel &intervals)
     {
-        int density = cg::utils::computeDensity(intervals);
+        int density = cg::interval_model_utils::computeDensity(intervals);
         cg::utils::Counters<PureOutputSensitive::Counts> counts;
         const auto& maybeMis = PureOutputSensitive::tryComputeMIS(intervals, density, counts); 
         if (maybeMis)
