+++ src/mis/distinct/valiente.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 #include <list>
 
-#include "data_structures/distinct_interval_rep.h"
+#include "data_structures/distinct_interval_model.h"
 #include "data_structures/interval.h"
 #include "mis/independent_set.h"
