+++ src/mis/shared/naive.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <set>
 
-#include "data_structures/shared_interval_rep.h"
+#include "data_structures/shared_interval_model.h"
 #include "data_structures/interval.h"
 #include "mis/independent_set.h"
 #include "utils/counters.h"
