+++ src/mis/shared/pruned_output_sensitive.cpp
@@ -3,7 +3,7 @@
 #include <list>
 #include <ranges>
 
-#include "data_structures/shared_interval_rep.h"
+#include "data_structures/shared_interval_model.h"
 #include "data_structures/interval.h"
 #include "mis/independent_set.h"
