+++ src/mis/shared/valiente.cpp
@@ -2,7 +2,7 @@
 #include <list>
 #include <algorithm>
 
-#include "data_structures/shared_interval_rep.h"
+#include "data_structures/shared_interval_model.h"
 #include "data_structures/interval.h"
 #include "mis/independent_set.h"
 #include "utils/counters.h"
